+++ product-search-api/src/main/java/org/ecommerce/productsearchapi/repository/ProductRepository.java
@@ -1,7 +1,7 @@
-package org.ecommerce.productsearchapi.repository.jpa;
+package org.ecommerce.productsearchapi.repository;
 
 import org.ecommerce.product.entity.Product;
-import org.ecommerce.productsearchapi.repository.ProductCustomRepository;
+import org.ecommerce.productsearchapi.repository.impl.ProductCustomRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
