+++ product-search-api/src/main/java/org/ecommerce/productsearchapi/repository/SellerRepository.java
@@ -1,4 +1,4 @@
-package org.ecommerce.productsearchapi.repository.jpa;
+package org.ecommerce.productsearchapi.repository;
 
 import org.ecommerce.product.entity.SellerRep;
 import org.springframework.data.jpa.repository.JpaRepository;
