+++ product-search-api/src/main/java/org/ecommerce/productsearchapi/service/ProductSearchService.java
@@ -3,7 +3,8 @@ package org.ecommerce.productsearchapi.service;
 import org.ecommerce.common.error.CustomException;
 import org.ecommerce.productsearchapi.dto.ProductSearchDto;
 import org.ecommerce.productsearchapi.exception.ProductSearchErrorCode;
-import org.ecommerce.productsearchapi.repository.querydsl.ProductCustomRepositoryImpl;
+import org.ecommerce.productsearchapi.repository.ProductRepository;
+import org.ecommerce.productsearchapi.repository.impl.ProductRepositoryImpl;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
@@ -14,7 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class ProductSearchService {
 
-	private final ProductCustomRepositoryImpl productCustomRepository;
+	private final ProductRepository productRepository;
 
 	/**
 	 * @apiNote 노원호
@@ -23,7 +24,7 @@ public class ProductSearchService {
 	 * @return ProductSearchDto
 	 */
 	public ProductSearchDto getProductById(final Integer productId) {
-		return productCustomRepository.findProductById(productId)
+		return productRepository.findProductById(productId)
 			.orElseThrow(() -> new CustomException(ProductSearchErrorCode.NOT_FOUND_PRODUCT_ID));
 	}
